+++ source/testing/CXXEnzymeExecutableModel.h
@@ -68,7 +68,7 @@ public:
     virtual int getNumFloatingSpecies();
 
     virtual int getFloatingSpeciesIndex(const std::string& eid);
-    virtual std::string getFloatingSpeciesId(int index);
+    virtual std::string getFloatingSpeciesId(size_t index);
 
     /**
      * independent species do are not defined by rules, they typically participate
@@ -175,7 +175,7 @@ public:
      */
     virtual int getNumBoundarySpecies();
     virtual int getBoundarySpeciesIndex(const std::string &eid);
-    virtual std::string getBoundarySpeciesId(int index);
+    virtual std::string getBoundarySpeciesId(size_t index);
 
     /**
      * get the boundary species amounts
@@ -234,7 +234,7 @@ public:
     /**
      * id of the indexed global parameter.
      */
-    virtual std::string getGlobalParameterId(int index);
+    virtual std::string getGlobalParameterId(size_t index);
 
     /**
      * get the global parameter values
@@ -276,9 +276,9 @@ public:
     /******************************************************************************/
 
     virtual int getNumCompartments();
-	virtual int getCompartmentIndexForFloatingSpecies(int index);
+	virtual int getCompartmentIndexForFloatingSpecies(size_t index);
     virtual int getCompartmentIndex(const std::string& eid);
-    virtual std::string getCompartmentId(int index);
+    virtual std::string getCompartmentId(size_t index);
 
     /**
      * get the compartment volumes
@@ -377,7 +377,7 @@ public:
 
     virtual int getNumConservedMoieties();
     virtual int getConservedMoietyIndex(const std::string& eid);
-    virtual std::string getConservedMoietyId(int index);
+    virtual std::string getConservedMoietyId(size_t index);
     virtual int getConservedMoietyValues(size_t len, int const *indx, double *values);
     virtual int setConservedMoietyValues(size_t len, int const *indx,
             const double *values);
@@ -398,7 +398,7 @@ public:
     /**
      * get the name of the specified reaction
      */
-    virtual std::string getReactionId(int index);
+    virtual std::string getReactionId(size_t index);
 
     /**
      * get the vector of reaction rates.
@@ -423,7 +423,7 @@ public:
     /**
      * get the id of an element of the state vector.
      */
-    virtual std::string getStateVectorId(int index);
+    virtual std::string getStateVectorId(size_t index);
 
     /**
      * The state vector is a vector of elements that are defined by
@@ -532,9 +532,9 @@ public:
      * If there is no event with this id, returns -1.
      */
     virtual int getEventIndex(const std::string& eid);
-    virtual std::string getEventId(int index);
-    virtual void setEventListener(int index, rr::EventListenerPtr eventHandler);
-    virtual rr::EventListenerPtr getEventListener(int index);
+    virtual std::string getEventId(size_t index);
+    virtual void setEventListener(size_t index, rr::EventListenerPtr eventHandler);
+    virtual rr::EventListenerPtr getEventListener(size_t index);
 
     /**
      * Get the amount rate of change for the i'th floating species
@@ -549,7 +549,7 @@ public:
      * @param index: index of the desired floating speceis rate.
      * @param reactionRates: pointer to buffer of reaction rates.
      */
-    virtual double getFloatingSpeciesAmountRate(int index,
+    virtual double getFloatingSpeciesAmountRate(size_t index,
             const double *reactionRates);
 
     /**
